+++ app/Http/Controllers/HomeController.php
@@ -417,42 +417,44 @@ class HomeController extends Controller
             $identitas_perusahaan->save();
 
             // Cabang
-            $cabang = Cabang::where('identitas_perusahaan_id', $dekripsi);
-            if($cabang->count() > 0) {
-                $cabang->delete();
-                // if(!isEmpty($request->nitku_cabang)) {
-                    if($request->nitku_cabang) {
-                        return ['status' => false, 'error' => 'NITKU cabang harus diisi'];
-                    }
-
-                    for($i = 0; $i < count($request->nitku_cabang); $i++) {
-                        Cabang::insert([
-                            'identitas_perusahaan_id' => $identitas_perusahaan->id,
-                            'nitku' => $request->nitku_cabang[$i],
-                            'nama' => $request->nama_cabang[$i],
-                            'alamat' => $request->alamat_nitku[$i],
-                            'created_at' => Carbon::now(),
-                            'updated_at' => Carbon::now(),
-                        ]);
-                    }
-                // }
-            } else {
-                // if(!isEmpty($request->nitku_cabang)) {
-                    if($request->nitku_cabang) {
-                        return ['status' => false, 'error' => 'NITKU cabang harus diisi'];
-                    }
-
-                    for($i = 0; $i < count($request->nitku_cabang); $i++) {
-                        Cabang::insert([
-                            'identitas_perusahaan_id' => $identitas_perusahaan->id,
-                            'nitku' => $request->nitku_cabang[$i],
-                            'nama' => $request->nama_cabang[$i],
-                            'alamat' => $request->alamat_nitku[$i],
-                            'created_at' => Carbon::now(),
-                            'updated_at' => Carbon::now(),
-                        ]);
-                    }
-                // }
+            if($request->bentuk_usaha == 'badan_usaha') {
+                $cabang = Cabang::where('identitas_perusahaan_id', $dekripsi);
+                if($cabang->count() > 0) {
+                    $cabang->delete();
+                    // if(!isEmpty($request->nitku_cabang)) {
+                        if($request->nitku_cabang) {
+                            return ['status' => false, 'error' => 'NITKU cabang harus diisi'];
+                        }
+    
+                        for($i = 0; $i < count($request->nitku_cabang); $i++) {
+                            Cabang::insert([
+                                'identitas_perusahaan_id' => $identitas_perusahaan->id,
+                                'nitku' => $request->nitku_cabang[$i],
+                                'nama' => $request->nama_cabang[$i],
+                                'alamat' => $request->alamat_nitku[$i],
+                                'created_at' => Carbon::now(),
+                                'updated_at' => Carbon::now(),
+                            ]);
+                        }
+                    // }
+                } else {
+                    // if(!isEmpty($request->nitku_cabang)) {
+                        if($request->nitku_cabang) {
+                            return ['status' => false, 'error' => 'NITKU cabang harus diisi'];
+                        }
+    
+                        for($i = 0; $i < count($request->nitku_cabang); $i++) {
+                            Cabang::insert([
+                                'identitas_perusahaan_id' => $identitas_perusahaan->id,
+                                'nitku' => $request->nitku_cabang[$i],
+                                'nama' => $request->nama_cabang[$i],
+                                'alamat' => $request->alamat_nitku[$i],
+                                'created_at' => Carbon::now(),
+                                'updated_at' => Carbon::now(),
+                            ]);
+                        }
+                    // }
+                }
             }
 
             // Identitas penanggung jawab
